Extract Direction class from Rover
Extract a new Direction class that takes a symbol corresponding to a cardinal point and responds to #right and #left to encapsulate rotation logic. Update the Rover class to call #right and #left on the new Direction object when turning left or right. Update CommandCentre to create a new Direction object from user input and inject this into the Rover object when deploying