Create individual Direction classes
Individual direction classes contain the logic for what direction is left or right respectively. The generic Direction class will act as a wrapper object that delegates to it's current instance of N, E, S or W. These classes will eventually also encapsulate the logic of how to move a coordinate forward with respect to the current direction or bearing.